Port StaxXmlPartialSchemaParser to InferSchema like json datasource.
