+++ src/main/scala/com/databricks/spark/xml/util/InferSchema.scala
@@ -15,10 +15,24 @@
  */
 package com.databricks.spark.xml.util
 
+import java.io.ByteArrayInputStream
+import javax.xml.stream.events._
+import javax.xml.stream.{XMLStreamException, XMLStreamConstants, XMLEventReader, XMLInputFactory}
+
+import org.slf4j.LoggerFactory
+
+import scala.collection.Seq
+import scala.collection.mutable.ArrayBuffer
+import scala.collection.JavaConversions._
+
 import org.apache.spark.rdd.RDD
 import org.apache.spark.sql.types._
+import com.databricks.spark.xml.util.TypeCast._
+import com.databricks.spark.xml.parsers.stax.StaxXmlParser._
+import com.databricks.spark.xml.parsers.stax.StaxConfiguration
 
 private[xml] object InferSchema {
+  private val logger = LoggerFactory.getLogger(InferSchema.getClass)
 
   /**
    * Copied from internal Spark api
@@ -54,10 +68,45 @@ private[xml] object InferSchema {
    *   2. Merge types by choosing the lowest type necessary to cover equal keys
    *   3. Replace any remaining null fields with string, the top type
    */
-  def infer(partialSchema: RDD[DataType]): StructType = {
+  def infer(xml: RDD[String],
+            conf: StaxConfiguration): StructType = {
+    require(conf.samplingRatio > 0, s"samplingRatio ($conf.samplingRatio) should be greater than 0")
+    val schemaData = if (conf.samplingRatio > 0.99) {
+      xml
+    } else {
+      xml.sample(withReplacement = false, conf.samplingRatio, 1)
+    }
+    val failFast = conf.failFastFlag
     // perform schema inference on each row and merge afterwards
-    val rootType = partialSchema
-      .treeAggregate[DataType](StructType(Seq()))(compatibleRootType, compatibleRootType)
+    val rootType = schemaData.mapPartitions { iter =>
+      iter.flatMap { xml =>
+        // It does not have to skip for white space, since [[XmlInputFormat]]
+        // always finds the root tag without a heading space.
+        val factory = XMLInputFactory.newInstance()
+        val reader = new ByteArrayInputStream(xml.getBytes)
+        val parser = factory.createXMLEventReader(reader)
+        try {
+          val rootEvent = skipUntil(parser, XMLStreamConstants.START_ELEMENT)
+          val rootAttributes = rootEvent.asStartElement.getAttributes
+            .map(_.asInstanceOf[Attribute]).toArray
+          Some(inferObject(parser, conf, rootAttributes))
+        } catch {
+          case _: java.lang.NumberFormatException | _: IllegalArgumentException if !failFast =>
+            logger.warn("Number format exception. " +
+              s"Dropping malformed line: ${xml.replaceAll("\n", "")}")
+            None
+          case _: java.text.ParseException if !failFast =>
+            logger.warn("Parse exception. " +
+              s"Dropping malformed line: ${xml.replaceAll("\n", "")}")
+            None
+          case _: XMLStreamException if !failFast =>
+            logger.warn(s"Dropping malformed row: ${xml.replaceAll("\n", "")}")
+            None
+          case _: XMLStreamException if failFast =>
+            throw new RuntimeException(s"Malformed row (failing fast): ${xml.replaceAll("\n", "")}")
+        }
+      }
+    }.treeAggregate[DataType](StructType(Seq()))(compatibleRootType, compatibleRootType)
 
     canonicalizeType(rootType) match {
       case Some(st: StructType) => st
@@ -65,6 +114,133 @@ private[xml] object InferSchema {
         // canonicalizeType erases all empty structs, including the only one we want to keep
         StructType(Seq())
     }
+  }
+
+  private def inferTypeFromString(value: String): DataType = {
+    if (Option(value).isEmpty) {
+      NullType
+    } else if (isLong(value)) {
+      LongType
+    } else if (isInteger(value)) {
+      IntegerType
+    } else if (isDouble(value)) {
+      DoubleType
+    } else if (isBoolean(value)) {
+      BooleanType
+    } else if (isTimestamp(value)) {
+      TimestampType
+    } else {
+      StringType
+    }
+  }
+
+  private def inferField(parser: XMLEventReader, conf: StaxConfiguration): DataType = {
+    val current = parser.peek
+    current match {
+      case _: EndElement =>
+        NullType
+      case _: StartElement =>
+        inferObject(parser, conf)
+      case c: Characters if !c.isIgnorableWhiteSpace && c.isWhiteSpace =>
+        // When `Characters` is found, we need to look further to decide
+        // if this is really data or space between other elements.
+        val next = {
+          parser.nextEvent
+          parser.peek
+        }
+        next match {
+          case _: EndElement =>
+            if (conf.treatEmptyValuesAsNulls) {
+              NullType
+            } else {
+              StringType
+            }
+          case _: StartElement =>
+            inferObject(parser, conf)
+        }
+      case c: Characters if !c.isIgnorableWhiteSpace && !c.isWhiteSpace =>
+        // This means data exists
+        inferTypeFromString(c.asCharacters().getData)
+
+      case e: XMLEvent =>
+        sys.error(s"Failed to parse data with unexpected event ${e.toString}")
+    }
+  }
+
+  /**
+   * Infer the type of a xml document from the parser's token stream
+   */
+  private def inferObject(parser: XMLEventReader,
+                          conf: StaxConfiguration,
+                          rootAttributes: Array[Attribute] = Array()): DataType = {
+    def toValuesMap(attributes: Array[Attribute]): Map[String, String] = {
+      if (conf.excludeAttributeFlag){
+        Map.empty[String, String]
+      } else {
+        val attrFields = attributes.map(conf.attributePrefix + _.getName.getLocalPart)
+        val attrDataTypes = attributes.map(_.getValue)
+        attrFields.zip(attrDataTypes).toMap
+      }
+    }
+
+    val builder = Seq.newBuilder[StructField]
+    val nameToDataTypes = collection.mutable.Map.empty[String, ArrayBuffer[DataType]]
+    var shouldStop = false
+    while (!shouldStop) {
+      parser.nextEvent match {
+        case e: StartElement =>
+          // If there are attributes, then we should process them first.
+          toValuesMap(rootAttributes).foreach {
+            case (f, v) =>
+              nameToDataTypes += (f -> ArrayBuffer(inferTypeFromString(v)))
+          }
+          val valuesMap = {
+            val attributes = e.getAttributes.map(_.asInstanceOf[Attribute]).toArray
+            toValuesMap(attributes)
+          }
+          val field = e.asStartElement.getName.getLocalPart
+          val inferredType = inferField(parser, conf)
+          inferredType match {
+            case st: StructType =>
+              val nestedBuilder = Seq.newBuilder[StructField]
+              nestedBuilder ++= st.fields
+              valuesMap.foreach {
+                case (f, v) =>
+                  nestedBuilder += StructField(f, inferTypeFromString(v), nullable = true)
+              }
+              val dataTypes = nameToDataTypes.getOrElse(field, ArrayBuffer.empty[DataType])
+              dataTypes += StructType(nestedBuilder.result().sortBy(_.name))
+              nameToDataTypes += (field -> dataTypes)
+            case _ if valuesMap.nonEmpty =>
+              // We need to wrap the attributes with a wrapper.
+              val nestedBuilder = Seq.newBuilder[StructField]
+              valuesMap.foreach {
+                case (f, v) =>
+                  nestedBuilder += StructField(f, inferTypeFromString(v), nullable = true)
+              }
+              nestedBuilder += StructField(conf.valueTag, inferredType, nullable = true)
+              nameToDataTypes +=
+                (field -> ArrayBuffer(StructType(nestedBuilder.result().sortBy(_.name))))
+            case _ =>
+              val dataTypes = nameToDataTypes.getOrElse(field, ArrayBuffer.empty[DataType])
+              dataTypes += inferredType
+              nameToDataTypes += (field -> dataTypes)
+          }
+        case _: EndElement =>
+          shouldStop = checkEndElement(parser, conf)
+        case _ =>
+          shouldStop = shouldStop && parser.hasNext
+      }
+    }
+    // We need to manually merges all the [[ArrayType]]s.
+    nameToDataTypes.foreach{
+      case (field, dataTypes) if dataTypes.length > 1 =>
+        val elementType = dataTypes.reduceLeft(InferSchema.compatibleType)
+        builder += StructField(field, ArrayType(elementType), nullable = true)
+      case (field, dataTypes) =>
+        builder += StructField(field, dataTypes.head, nullable = true)
+    }
+    StructType(builder.result().sortBy(_.name))
   }
 
   /**
